Ref: Change int to double in weightSize attribute

+++ src/main/java/sizebay/catalog/client/model/ProductBasicInformation.java
@@ -39,5 +39,5 @@ public class ProductBasicInformation implements Serializable {
 	String strongSubcategoryName;
 	int strongModelId;
 	String strongModelName;
-	int weightSize;
+	double weightSize;
 }
